third_party/blink/web_tests/platform/mac-mac10.13/external/wpt/screen-capture/permissions-policy-audio+video.https.sub-expected.txt

This is a testharness.js-based test.
PASS Default "display-capture" permissions policy ["self"] allows the top-level document.
PASS Default "display-capture" permissions policy ["self"] allows same-origin iframes.
FAIL Default "display-capture" permissions policy ["self"] disallows cross-origin iframes. assert_equals: element click intercepted error. expected "NotAllowedError" but got "Error"
PASS permissions policy "display-capture" can be enabled in cross-origin iframes using "allow" attribute.
PASS permissions policy "display-capture" can be disabled in same-origin iframes using "allow" attribute.
Harness: the test ran to completion.
